Unkn library and initial code for esp8266(master) added
the library has drive() function that gives power to some pins to make the motors rotate. It also has the function init() to initialize the motors and it has MAX defined as 255 for maximum power. It has dance.ino, a very simple example you can start with.

The arduino and the esp are connected through SPI for a variety of reasons that will be very soon explained in the readme.md. Esp is the master, arduino is the only slave.